Saturation to negative values of the membrane
My experiments with a simplified version of the simulator are showing
that the membrane behaviour is a lot more complex when there’s a
saturation to the negative values too.
I also changed the way the spikes are updated. Now the system tests the
threshold only in the next step (not at the same step the membrane
reached it).