Fix for a sequence of wrong commits
The main C code was changed to allow reset voltage values to be drawn
from a distribution. But then I saw the example was not working and, by
mistake, reverted to an old version instead of the fixed one. The only
thing I should have done was to fix the example. The example was fixed
by enabling all neurons to have their membrane value updated (by
default only neurons with at least one connection are updated) and also
by setting the reset voltage value to the same as the Brian simulation.